Added a link to the API specification
This link is based directly on the ApiDocumentation.yml file in this repository so any changes should be reflected by a simple refresh of the page.
Another benefit is that this view is reader only (and hence has greater space than the usual view with the editor).

p.s. since it's simple and small (and not concurrent with other changes), this will be committed directly to the main branch.